--- src/const/index.ts
@@ -1,4 +1,0 @@
-export { ColorMap } from '~/const/colorMap';
-export { PieceColor } from '~/const/pieceColorMap';
-export { PlayerMap } from '~/const/playerMap';
-export { SizeMap } from '~/const/sizeMap';
