+++ src/constants/colorMap.ts
