+++ src/constants/index.ts
@@ -1,0 +1,4 @@
+export { ColorMap } from '~/constants/colorMap';
+export { PieceColor } from '~/constants/pieceColorMap';
+export { PlayerMap } from '~/constants/playerMap';
+export { SizeMap } from '~/constants/sizeMap';
