+++ src/constants/pieceColorMap.ts
